Migrating the interpreter class and accessories to the model library.
