+++ core/CMakeLists.txt
@@ -14,13 +14,12 @@ add_polymec_library(polymec_core polymec.c arch.c allocators.c logging.c timer.c
                     special_functions.c point.c point2.c tensor2.c tagger.c mesh.c 
                     point_cloud.c hilbert.c kd_tree.c octree.c rng.c #parallel_qsort.c
                     permutations.c adj_graph.c sp_func.c st_func.c polynomial.c 
-                    least_squares.c index_space.c newton.c options.c interpreter.c 
+                    least_squares.c index_space.c newton.c options.c  
                     mesh_node_exchangers.c partition_mesh.c partition_point_cloud.c
-                    text_buffer.c serializer.c interpreter_register_default_functions.c 
-                    thread_pool.c string_utils.c file_utils.c array_utils.c lookup1.c
-                    linear_algebra.c local_matrix.c block_diagonal_matrix.c
-                    sparse_local_matrix.c dense_local_matrix ${DYLOAD_SOURCES}
-                    silo_file.c exchanger.c)
+                    text_buffer.c serializer.c thread_pool.c string_utils.c 
+                    file_utils.c array_utils.c lookup1.c linear_algebra.c 
+                    local_matrix.c block_diagonal_matrix.c sparse_local_matrix.c 
+                    dense_local_matrix ${DYLOAD_SOURCES} silo_file.c exchanger.c)
 
 # polymec.c depends on the creation of polymec_version.h.
 add_dependencies(polymec_core update_version_h)
