+++ geometry/CMakeLists.txt
@@ -1,6 +1,5 @@
 set(MESH_GENERATION_SOURCES create_boundary_generators.c 
-    generate_random_points.c point_factory.c create_point_lattice.c
-    mesh_factory.c cartesian_1d.c cylindrical_1d.c spherical_1d.c
+    generate_random_points.c create_point_lattice.c
     create_welded_block_mesh.c rotate_mesh.c translate_mesh.c
     create_cubed_cylinder_mesh.c create_cubed_sphere_mesh.c
     delaunay_triangulation.c create_convex_hull.c create_rectilinear_mesh.c 
@@ -12,8 +11,7 @@ add_polymec_library(polymec_geometry coord_mapping.c plane.c cylinder.c sphere.c
                     intersection.c union.c difference.c scaled.c rect_prism.c 
                     giftwrap_hull.c polygon2.c polygon.c tetrahedron.c 
                     ${MESH_GENERATION_SOURCES} cubic_lattice.c
-                    symmetry_1d.c interpreter_register_geometry_functions.c 
-                    interpreter_register_spfuncs.c)
+                    symmetry_1d.c)
 
 set(POLYMEC_LIBRARIES polymec_geometry;${POLYMEC_LIBRARIES} PARENT_SCOPE)
 set(POLYMEC_LIBRARIES polymec_geometry;${POLYMEC_LIBRARIES})
