--- geometry/interpreter_register_geometry_functions.h
@@ -1,18 +1,0 @@
-// Copyright (c) 2012-2016, Jeffrey N. Johnson
-// All rights reserved.
-// 
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-#ifndef POLYMEC_INTERPRETER_REGISTER_GEOMETRY_FUNCTIONS_H
-#define POLYMEC_INTERPRETER_REGISTER_GEOMETRY_FUNCTIONS_H
-
-#include "core/interpreter.h"
-
-// This function registers basic geometry-related functions for the
-// interpreter.
-void interpreter_register_geometry_functions(interpreter_t* interp);
-
-#endif
-
