+++ model/CMakeLists.txt
@@ -2,7 +2,13 @@ include(add_polymec_library)
 add_polymec_library(polymec_model model.c boundary_cell_map.c periodic_bc.c 
                     point_weight_function.c stencil.c mesh_stencils.c 
                     neighbor_pairing.c polynomial_fit.c point_spacing_estimator.c
-                    aux_state.c physical_constants.c interpreter_register_model_functions.c)
+                    aux_state.c physical_constants.c interpreter.c
+                    interpreter_register_default_functions.c 
+                    interpreter_register_geometry_functions.c
+                    interpreter_register_spfuncs.c
+                    interpreter_register_model_functions.c
+                    mesh_factory.c point_factory.c
+                    cartesian_1d.c cylindrical_1d.c spherical_1d.c)
 
 set(POLYMEC_LIBRARIES polymec_model;${POLYMEC_LIBRARIES} PARENT_SCOPE)
 set(POLYMEC_LIBRARIES polymec_model;${POLYMEC_LIBRARIES})
