+++ model/cylindrical_1d.c
@@ -5,27 +5,27 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-// spherical_1d.c - Implementations of interpreter functions for generating
-// 1D spherical meshes.
+// cylindrical_1d.c - Implementations of interpreter functions for generating
+// 1D cylindrical meshes.
 
 #include <strings.h>
 #include "core/mesh.h"
-#include "core/interpreter.h"
 #include "geometry/symmetry.h"
+#include "model/interpreter.h"
 
 // Lua stuff.
 #include "lua.h"
 #include "lualib.h"
 #include "lauxlib.h"
 
-int spherical_1d_uniform(lua_State* lua)
+int cylindrical_1d_uniform(lua_State* lua)
 {
   // Check the arguments.
   int num_args = lua_gettop(lua);
   if (num_args != 3)
   {
     return luaL_error(lua, "Invalid arguments. Usage:\n"
-                      "mesh = spherical_1d.uniform(r1, r2, N)");
+                      "mesh = cylindrical_1d.uniform(r1, r2, N)");
   }
 
   // Get the arguments.
@@ -40,7 +40,7 @@ int spherical_1d_uniform(lua_State* lua)
     return luaL_error(lua, "N must be positive.");
 
   // Create the mesh.
-  mesh_t* mesh = create_uniform_spherical_1d_mesh(MPI_COMM_WORLD, r1, r2, N);
+  mesh_t* mesh = create_uniform_cylindrical_1d_mesh(MPI_COMM_WORLD, r1, r2, N);
 
   // Push the mesh onto the stack.
   lua_pushmesh(lua, mesh);
@@ -48,14 +48,14 @@ int spherical_1d_uniform(lua_State* lua)
   return 1;
 }
 
-int spherical_1d_logarithmic(lua_State* lua)
+int cylindrical_1d_logarithmic(lua_State* lua)
 {
   // Check the arguments.
   int num_args = lua_gettop(lua);
   if (num_args != 4)
   {
     return luaL_error(lua, "Invalid arguments. Usage:\n"
-                      "mesh = spherical_1d.logarithmic(r1, r2, log_factor, N)");
+                      "mesh = cylindrical_1d.logarithmic(r1, r2, log_factor, N)");
   }
 
   // Get the arguments.
@@ -68,12 +68,12 @@ int spherical_1d_logarithmic(lua_State* lua)
   real_t log_factor = (real_t)lua_tonumber(lua, 3);
   if (log_factor <= 0.0)
     return luaL_error(lua, "log factor must be positive.");
-  int N = (int)lua_tonumber(lua, 4);
+  int N = (int)lua_tonumber(lua, 3);
   if (N <= 0)
     return luaL_error(lua, "N must be positive.");
 
   // Create the mesh.
-  mesh_t* mesh = create_logarithmic_spherical_1d_mesh(MPI_COMM_WORLD, r1, r2, log_factor, N);
+  mesh_t* mesh = create_logarithmic_cylindrical_1d_mesh(MPI_COMM_WORLD, r1, r2, log_factor, N);
 
   // Push the mesh onto the stack.
   lua_pushmesh(lua, mesh);
@@ -81,14 +81,14 @@ int spherical_1d_logarithmic(lua_State* lua)
   return 1;
 }
 
-int spherical_1d_nonuniform(lua_State* lua)
+int cylindrical_1d_nonuniform(lua_State* lua)
 {
   // Check the arguments.
   int num_args = lua_gettop(lua);
   if (num_args != 1)
   {
     return luaL_error(lua, "Invalid arguments. Usage:\n"
-                      "mesh = spherical_1d.nonuniform({r1, r2, ..., rN})");
+                      "mesh = cylindrical_1d.nonuniform({r1, r2, ..., rN})");
   }
 
   // Get the arguments.
@@ -105,7 +105,7 @@ int spherical_1d_nonuniform(lua_State* lua)
   }
 
   // Create the mesh.
-  mesh_t* mesh = create_nonuniform_spherical_1d_mesh(MPI_COMM_WORLD, rs, N);
+  mesh_t* mesh = create_nonuniform_cylindrical_1d_mesh(MPI_COMM_WORLD, rs, N);
 
   // Push the mesh onto the stack.
   lua_pushmesh(lua, mesh);
