+++ model/interpreter_register_model_functions.c
@@ -5,7 +5,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "model/interpreter_register_model_functions.h"
+#include "model/interpreter.h"
 #include "model/periodic_bc.h"
 
 // Lua stuff.
