+++ model/interpreter_register_spfuncs.c
@@ -5,9 +5,9 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "core/interpreter.h"
 #include "geometry/sphere.h"
 #include "geometry/rect_prism.h"
+#include "model/interpreter.h"
 
 // Lua stuff.
 #include "lua.h"
